Starting application upgrade
- Upgrade npm packages to latest versions
- Fix console errors coming from new browser APIs (specifically relates to
  audio and video)
- Clean up console warnings
- Starting the process of taking out API calls to Azure and IBM Watson
