--- frontend/public/base64-binary.js
@@ -1,176 +1,0 @@
-/*
-Copyright (c) 2011, Daniel Guerrero
-All rights reserved.
-
-Redistribution and use in source and binary forms, with or without
-modification, are permitted provided that the following conditions are met:
-    * Redistributions of source code must retain the above copyright
-      notice, this list of conditions and the following disclaimer.
-    * Redistributions in binary form must reproduce the above copyright
-      notice, this list of conditions and the following disclaimer in the
-      documentation and/or other materials provided with the distribution.
-
-THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
-ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-DISCLAIMED. IN NO EVENT SHALL DANIEL GUERRERO BE LIABLE FOR ANY
-DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-(INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
-ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-/**
- * Uses the new array typed in javascript to binary base64 encode/decode
- * at the moment just decodes a binary base64 encoded
- * into either an ArrayBuffer (decodeArrayBuffer)
- * or into an Uint8Array (decode)
- * 
- * References:
- * https://developer.mozilla.org/en/JavaScript_typed_arrays/ArrayBuffer
- * https://developer.mozilla.org/en/JavaScript_typed_arrays/Uint8Array
- */
-
-var sessionTime;
-
-var globalData = [{ 
-	"key" : "Anger" ,
-	"values" : [ ]
-}, 
-
-{ 
-	"key" : "Contempt",
-	"values" : [ ]
-}, 
-
-{ 
-	"key" : "Disgust",
-	"values" : [ ]
-},
-
-{ 
-	"key" : "Fear",
-	"values" : [ ]
-} , 
-
-{ 
-	"key" : "Happiness",
-	"values" : [ ]
-} , 
-
-{
-	"key" : "Neutral",
-	"values" : [ ]
-} , 
-
-{ 
-	"key" : "Sadness",
-	"values" : [ ]
-} ,
-{ 
-	"key" : "Surprise",
-	"values" : [ ]
-}
-];
-var localData = [ 
-	{ 
-		"key" : "Anger" , 
-		"values" : [ [0, 0] , [1, 0] , [2, 0] , [3, 0] , [4, 0] , [5, 0] , [6, 0] , [7, 0] , [8, 0] , [9, 0] , [10, 0] , [11, 0] , [12, 0] , [13, 0] , [14, 0] , [15, 0] , [16, 0] , [17, 0] , [18, 0] , [19, 0] , [20, 0] , [21, 0] , [22, 0] , [23, 0] , [24, 0] , [25, 0] , [26, 0] , [27, 0] , [28, 0] , [29, 0] , [30, 0] , [31, 0] , [32, 0] , [33, 0] , [34, 0] , [35, 0] , [36, 0] , [37, 0] , [38, 0] , [39, 0] , [40, 0] , [41, 0] , [42, 0] , [43, 0] , [44, 0] , [45, 0] , [46, 0] , [47, 0] , [48, 0] , [49, 0] , [50, 0] , [51, 0] , [52, 0] , [53, 0] , [54, 0] , [55, 0] , [56, 0] , [57, 0] , [58, 0] , [59, 0] , [60, 0] , [61, 0] , [62, 0] , [63, 0] , [64, 0] , [65, 0] , [66, 0] , [67, 0] , [68, 0] , [69, 0] , [70, 0] , [71, 0] , [72, 0] , [73, 0] , [74, 0] , [75, 0] , [76, 0] , [77, 0] , [78, 0] , [79, 0] , [80, 0] , [81, 0] , [82, 0] , [83, 0] , [84, 0] , [85, 0] , [86, 0] , [87, 0] , [88, 0] , [89, 0] , [90, 0] , [91, 0] , [92, 0] , [93, 0] , [94, 0] , [95, 0] , [96, 0] , [97, 0] , [98, 0] , [99, 0] ]
-	}, 
-
-	{ 
-		"key" : "Contempt" , 
-		"values" : [ [0, 0] , [1, 0] , [2, 0] , [3, 0] , [4, 0] , [5, 0] , [6, 0] , [7, 0] , [8, 0] , [9, 0] , [10, 0] , [11, 0] , [12, 0] , [13, 0] , [14, 0] , [15, 0] , [16, 0] , [17, 0] , [18, 0] , [19, 0] , [20, 0] , [21, 0] , [22, 0] , [23, 0] , [24, 0] , [25, 0] , [26, 0] , [27, 0] , [28, 0] , [29, 0] , [30, 0] , [31, 0] , [32, 0] , [33, 0] , [34, 0] , [35, 0] , [36, 0] , [37, 0] , [38, 0] , [39, 0] , [40, 0] , [41, 0] , [42, 0] , [43, 0] , [44, 0] , [45, 0] , [46, 0] , [47, 0] , [48, 0] , [49, 0] , [50, 0] , [51, 0] , [52, 0] , [53, 0] , [54, 0] , [55, 0] , [56, 0] , [57, 0] , [58, 0] , [59, 0] , [60, 0] , [61, 0] , [62, 0] , [63, 0] , [64, 0] , [65, 0] , [66, 0] , [67, 0] , [68, 0] , [69, 0] , [70, 0] , [71, 0] , [72, 0] , [73, 0] , [74, 0] , [75, 0] , [76, 0] , [77, 0] , [78, 0] , [79, 0] , [80, 0] , [81, 0] , [82, 0] , [83, 0] , [84, 0] , [85, 0] , [86, 0] , [87, 0] , [88, 0] , [89, 0] , [90, 0] , [91, 0] , [92, 0] , [93, 0] , [94, 0] , [95, 0] , [96, 0] , [97, 0] , [98, 0] , [99, 0]  ]
-	}, 
-
-	{ 
-		"key" : "Disgust" , 
-		"values" : [ [0, 0] , [1, 0] , [2, 0] , [3, 0] , [4, 0] , [5, 0] , [6, 0] , [7, 0] , [8, 0] , [9, 0] , [10, 0] , [11, 0] , [12, 0] , [13, 0] , [14, 0] , [15, 0] , [16, 0] , [17, 0] , [18, 0] , [19, 0] , [20, 0] , [21, 0] , [22, 0] , [23, 0] , [24, 0] , [25, 0] , [26, 0] , [27, 0] , [28, 0] , [29, 0] , [30, 0] , [31, 0] , [32, 0] , [33, 0] , [34, 0] , [35, 0] , [36, 0] , [37, 0] , [38, 0] , [39, 0] , [40, 0] , [41, 0] , [42, 0] , [43, 0] , [44, 0] , [45, 0] , [46, 0] , [47, 0] , [48, 0] , [49, 0] , [50, 0] , [51, 0] , [52, 0] , [53, 0] , [54, 0] , [55, 0] , [56, 0] , [57, 0] , [58, 0] , [59, 0] , [60, 0] , [61, 0] , [62, 0] , [63, 0] , [64, 0] , [65, 0] , [66, 0] , [67, 0] , [68, 0] , [69, 0] , [70, 0] , [71, 0] , [72, 0] , [73, 0] , [74, 0] , [75, 0] , [76, 0] , [77, 0] , [78, 0] , [79, 0] , [80, 0] , [81, 0] , [82, 0] , [83, 0] , [84, 0] , [85, 0] , [86, 0] , [87, 0] , [88, 0] , [89, 0] , [90, 0] , [91, 0] , [92, 0] , [93, 0] , [94, 0] , [95, 0] , [96, 0] , [97, 0] , [98, 0] , [99, 0] ]
-	},
-
-	{ 
-		"key" : "Fear" , 
-		"values" : [ [0, 0] , [1, 0] , [2, 0] , [3, 0] , [4, 0] , [5, 0] , [6, 0] , [7, 0] , [8, 0] , [9, 0] , [10, 0] , [11, 0] , [12, 0] , [13, 0] , [14, 0] , [15, 0] , [16, 0] , [17, 0] , [18, 0] , [19, 0] , [20, 0] , [21, 0] , [22, 0] , [23, 0] , [24, 0] , [25, 0] , [26, 0] , [27, 0] , [28, 0] , [29, 0] , [30, 0] , [31, 0] , [32, 0] , [33, 0] , [34, 0] , [35, 0] , [36, 0] , [37, 0] , [38, 0] , [39, 0] , [40, 0] , [41, 0] , [42, 0] , [43, 0] , [44, 0] , [45, 0] , [46, 0] , [47, 0] , [48, 0] , [49, 0] , [50, 0] , [51, 0] , [52, 0] , [53, 0] , [54, 0] , [55, 0] , [56, 0] , [57, 0] , [58, 0] , [59, 0] , [60, 0] , [61, 0] , [62, 0] , [63, 0] , [64, 0] , [65, 0] , [66, 0] , [67, 0] , [68, 0] , [69, 0] , [70, 0] , [71, 0] , [72, 0] , [73, 0] , [74, 0] , [75, 0] , [76, 0] , [77, 0] , [78, 0] , [79, 0] , [80, 0] , [81, 0] , [82, 0] , [83, 0] , [84, 0] , [85, 0] , [86, 0] , [87, 0] , [88, 0] , [89, 0] , [90, 0] , [91, 0] , [92, 0] , [93, 0] , [94, 0] , [95, 0] , [96, 0] , [97, 0] , [98, 0] , [99, 0] ]
-	} , 
-
-	{ 
-		"key" : "Happiness" , 
-		"values" : [ [0, 0] , [1, 0] , [2, 0] , [3, 0] , [4, 0] , [5, 0] , [6, 0] , [7, 0] , [8, 0] , [9, 0] , [10, 0] , [11, 0] , [12, 0] , [13, 0] , [14, 0] , [15, 0] , [16, 0] , [17, 0] , [18, 0] , [19, 0] , [20, 0] , [21, 0] , [22, 0] , [23, 0] , [24, 0] , [25, 0] , [26, 0] , [27, 0] , [28, 0] , [29, 0] , [30, 0] , [31, 0] , [32, 0] , [33, 0] , [34, 0] , [35, 0] , [36, 0] , [37, 0] , [38, 0] , [39, 0] , [40, 0] , [41, 0] , [42, 0] , [43, 0] , [44, 0] , [45, 0] , [46, 0] , [47, 0] , [48, 0] , [49, 0] , [50, 0] , [51, 0] , [52, 0] , [53, 0] , [54, 0] , [55, 0] , [56, 0] , [57, 0] , [58, 0] , [59, 0] , [60, 0] , [61, 0] , [62, 0] , [63, 0] , [64, 0] , [65, 0] , [66, 0] , [67, 0] , [68, 0] , [69, 0] , [70, 0] , [71, 0] , [72, 0] , [73, 0] , [74, 0] , [75, 0] , [76, 0] , [77, 0] , [78, 0] , [79, 0] , [80, 0] , [81, 0] , [82, 0] , [83, 0] , [84, 0] , [85, 0] , [86, 0] , [87, 0] , [88, 0] , [89, 0] , [90, 0] , [91, 0] , [92, 0] , [93, 0] , [94, 0] , [95, 0] , [96, 0] , [97, 0] , [98, 0] , [99, 0] ]
-	} , 
-
-	{
-		"key" : "Neutral" , 
-		"values" : [ [0, 0] , [1, 0] , [2, 0] , [3, 0] , [4, 0] , [5, 0] , [6, 0] , [7, 0] , [8, 0] , [9, 0] , [10, 0] , [11, 0] , [12, 0] , [13, 0] , [14, 0] , [15, 0] , [16, 0] , [17, 0] , [18, 0] , [19, 0] , [20, 0] , [21, 0] , [22, 0] , [23, 0] , [24, 0] , [25, 0] , [26, 0] , [27, 0] , [28, 0] , [29, 0] , [30, 0] , [31, 0] , [32, 0] , [33, 0] , [34, 0] , [35, 0] , [36, 0] , [37, 0] , [38, 0] , [39, 0] , [40, 0] , [41, 0] , [42, 0] , [43, 0] , [44, 0] , [45, 0] , [46, 0] , [47, 0] , [48, 0] , [49, 0] , [50, 0] , [51, 0] , [52, 0] , [53, 0] , [54, 0] , [55, 0] , [56, 0] , [57, 0] , [58, 0] , [59, 0] , [60, 0] , [61, 0] , [62, 0] , [63, 0] , [64, 0] , [65, 0] , [66, 0] , [67, 0] , [68, 0] , [69, 0] , [70, 0] , [71, 0] , [72, 0] , [73, 0] , [74, 0] , [75, 0] , [76, 0] , [77, 0] , [78, 0] , [79, 0] , [80, 0] , [81, 0] , [82, 0] , [83, 0] , [84, 0] , [85, 0] , [86, 0] , [87, 0] , [88, 0] , [89, 0] , [90, 0] , [91, 0] , [92, 0] , [93, 0] , [94, 0] , [95, 0] , [96, 0] , [97, 0] , [98, 0] , [99, 0]  ]
-	} , 
-
-	{ 
-		"key" : "Sadness" , 
-		"values" : [ [0, 0] , [1, 0] , [2, 0] , [3, 0] , [4, 0] , [5, 0] , [6, 0] , [7, 0] , [8, 0] , [9, 0] , [10, 0] , [11, 0] , [12, 0] , [13, 0] , [14, 0] , [15, 0] , [16, 0] , [17, 0] , [18, 0] , [19, 0] , [20, 0] , [21, 0] , [22, 0] , [23, 0] , [24, 0] , [25, 0] , [26, 0] , [27, 0] , [28, 0] , [29, 0] , [30, 0] , [31, 0] , [32, 0] , [33, 0] , [34, 0] , [35, 0] , [36, 0] , [37, 0] , [38, 0] , [39, 0] , [40, 0] , [41, 0] , [42, 0] , [43, 0] , [44, 0] , [45, 0] , [46, 0] , [47, 0] , [48, 0] , [49, 0] , [50, 0] , [51, 0] , [52, 0] , [53, 0] , [54, 0] , [55, 0] , [56, 0] , [57, 0] , [58, 0] , [59, 0] , [60, 0] , [61, 0] , [62, 0] , [63, 0] , [64, 0] , [65, 0] , [66, 0] , [67, 0] , [68, 0] , [69, 0] , [70, 0] , [71, 0] , [72, 0] , [73, 0] , [74, 0] , [75, 0] , [76, 0] , [77, 0] , [78, 0] , [79, 0] , [80, 0] , [81, 0] , [82, 0] , [83, 0] , [84, 0] , [85, 0] , [86, 0] , [87, 0] , [88, 0] , [89, 0] , [90, 0] , [91, 0] , [92, 0] , [93, 0] , [94, 0] , [95, 0] , [96, 0] , [97, 0] , [98, 0] , [99, 0]  ]
-	} ,
-	{ 
-		"key" : "Surprise" , 
-		"values" : [ [0, 0] , [1, 0] , [2, 0] , [3, 0] , [4, 0] , [5, 0] , [6, 0] , [7, 0] , [8, 0] , [9, 0] , [10, 0] , [11, 0] , [12, 0] , [13, 0] , [14, 0] , [15, 0] , [16, 0] , [17, 0] , [18, 0] , [19, 0] , [20, 0] , [21, 0] , [22, 0] , [23, 0] , [24, 0] , [25, 0] , [26, 0] , [27, 0] , [28, 0] , [29, 0] , [30, 0] , [31, 0] , [32, 0] , [33, 0] , [34, 0] , [35, 0] , [36, 0] , [37, 0] , [38, 0] , [39, 0] , [40, 0] , [41, 0] , [42, 0] , [43, 0] , [44, 0] , [45, 0] , [46, 0] , [47, 0] , [48, 0] , [49, 0] , [50, 0] , [51, 0] , [52, 0] , [53, 0] , [54, 0] , [55, 0] , [56, 0] , [57, 0] , [58, 0] , [59, 0] , [60, 0] , [61, 0] , [62, 0] , [63, 0] , [64, 0] , [65, 0] , [66, 0] , [67, 0] , [68, 0] , [69, 0] , [70, 0] , [71, 0] , [72, 0] , [73, 0] , [74, 0] , [75, 0] , [76, 0] , [77, 0] , [78, 0] , [79, 0] , [80, 0] , [81, 0] , [82, 0] , [83, 0] , [84, 0] , [85, 0] , [86, 0] , [87, 0] , [88, 0] , [89, 0] , [90, 0] , [91, 0] , [92, 0] , [93, 0] , [94, 0] , [95, 0] , [96, 0] , [97, 0] , [98, 0] , [99, 0]  ]
-	}
-];
-
-var Base64Binary = {
-	_keyStr : "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/=",
-	
-	/* will return a  Uint8Array type */
-	decodeArrayBuffer: function(input) {
-		var bytes = (input.length/4) * 3;
-		var ab = new ArrayBuffer(bytes);
-		this.decode(input, ab);
-		
-		return ab;
-	},
-
-	removePaddingChars: function(input){
-		var lkey = this._keyStr.indexOf(input.charAt(input.length - 1));
-		if(lkey == 64){
-			return input.substring(0,input.length - 1);
-		}
-		return input;
-	},
-
-	decode: function (input, arrayBuffer) {
-		//get last chars to see if are valid
-		input = this.removePaddingChars(input);
-		input = this.removePaddingChars(input);
-
-		var bytes = parseInt((input.length / 4) * 3, 10);
-		
-		var uarray;
-		var chr1, chr2, chr3;
-		var enc1, enc2, enc3, enc4;
-		var i = 0;
-		var j = 0;
-		
-		if (arrayBuffer)
-			uarray = new Uint8Array(arrayBuffer);
-		else
-			uarray = new Uint8Array(bytes);
-		
-		input = input.replace(/[^A-Za-z0-9\+\/\=]/g, "");
-		
-		for (i=0; i<bytes; i+=3) {	
-			//get the 3 octects in 4 ascii chars
-			enc1 = this._keyStr.indexOf(input.charAt(j++));
-			enc2 = this._keyStr.indexOf(input.charAt(j++));
-			enc3 = this._keyStr.indexOf(input.charAt(j++));
-			enc4 = this._keyStr.indexOf(input.charAt(j++));
-	
-			chr1 = (enc1 << 2) | (enc2 >> 4);
-			chr2 = ((enc2 & 15) << 4) | (enc3 >> 2);
-			chr3 = ((enc3 & 3) << 6) | enc4;
-	
-			uarray[i] = chr1;			
-			if (enc3 != 64) uarray[i+1] = chr2;
-			if (enc4 != 64) uarray[i+2] = chr3;
-		}
-	
-		return uarray;	
-	}
-}
