+++ src/App/App.js
@@ -1,50 +1,49 @@
 import React, { Component } from 'react'
-import { Button } from 'react-bootstrap'
 import Live from '../Pages/Live/Live'
 import Home from '../Pages/Home/Home'
 import Result from '../Pages/Result/Result'
 
 class App extends Component {
-  
+
   constructor() {
     super()
     this.state = {
       currentPage: 'Home',
-      historyData: [{ 
+      historyData: [{
           "key" : "Anger" ,
           "values" : [ ]
-        }, 
-    
-        { 
+        },
+
+        {
           "key" : "Contempt",
           "values" : [ ]
-        }, 
-    
-        { 
+        },
+
+        {
           "key" : "Disgust",
           "values" : [ ]
         },
-    
-        { 
+
+        {
           "key" : "Fear",
           "values" : [ ]
-        } , 
-    
-        { 
+        } ,
+
+        {
           "key" : "Happiness",
           "values" : [ ]
-        } , 
-    
+        } ,
+
         {
           "key" : "Neutral",
           "values" : [ ]
-        } , 
-    
-        { 
+        } ,
+
+        {
           "key" : "Sadness",
           "values" : [ ]
         } ,
-        { 
+        {
           "key" : "Surprise",
           "values" : [ ]
         }
@@ -62,7 +61,7 @@ class App extends Component {
     if (this.state.currentPage === 'live') {
       return <Live changePage={(pageName) => this.changePage(pageName)} />
     } else if (this.state.currentPage === 'result') {
-      return <Result changePage={(pageName) => this.changePage(pageName)} />  
+      return <Result changePage={(pageName) => this.changePage(pageName)} />
     }
     return <Home changePage={pageName => this.changePage(pageName)} />
   }
